+++ src/graph_onedrive/_main.py
@@ -2,6 +2,7 @@
 """
 from __future__ import annotations
 
+import logging
 import warnings
 from contextlib import contextmanager
 from pathlib import Path
@@ -22,8 +23,10 @@ def OneDriveManager(
     Returns:
         onedrive_instance (OneDrive) -- OneDrive object instance
     """
+    logging.info("OneDriveManager creating instance")
     onedrive_instance = OneDrive.from_json(file_path, config_key)
     yield onedrive_instance
+    logging.info("OneDriveManager saving instance configuration to file")
     onedrive_instance.to_json(file_path, config_key)
 
 
@@ -43,13 +46,14 @@ def create(
         client_id (str) -- Azure app client id
         client_secret (str) -- Azure app client secret
     Keyword arguments:
-        tenant (str) -- Azure app org tentent id number, use default if multi-tenent (default = "common")
+        tenant (str) -- Azure app org tentent id number, use default if multi-tenant (default = "common")
         redirect_url (str)  -- Authentication redirection url (default = "http://localhost:8080")
         refresh_token (str) -- optional token from previous session (default = None)
     Returns:
         onedrive_instance (OneDrive) -- OneDrive object instance
     """
     # Warn to use class directly
+    logging.warning("create() depreciated, use OneDrive()")
     warnings.warn(
         "create() depreciated, use OneDrive()",
         category=DeprecationWarning,
@@ -78,6 +82,7 @@ def create_from_config_file(
         onedrive_instance (OneDrive) -- OneDrive object instance
     """
     # Warn to use class directly
+    logging.warning("create_from_config_file() depreciated, use OneDrive.from_json()")
     warnings.warn(
         "create_from_config_file() depreciated, use OneDrive.from_json()",
         category=DeprecationWarning,
@@ -106,6 +111,7 @@ def save_to_config_file(
             f"onedrive_instance expected 'OneDrive', got {type(onedrive_instance).__name__!r}"
         )
     # Warn to use class directly
+    logging.warning("save_to_config_file() depreciated, use OneDrive.to_json()")
     warnings.warn(
         "save_to_config_file() depreciated, use OneDrive.to_json()",
         category=DeprecationWarning,
@@ -113,6 +119,3 @@ def save_to_config_file(
     )
     # Save to json
     onedrive_instance.to_json(config_path, config_key)
-
-    # Nothing returned which signals no errors
-    return